refactor(web): layout compartilhado das páginas
