+++ web/src/app/layout.tsx
@@ -5,6 +5,13 @@ import {
   Bai_Jamjuree as BaiJamjuree,
 } from 'next/font/google'
 
+import { Copyright } from '@/components/Copyright'
+import { Hero } from '@/components/Hero'
+import { SignIn } from '@/components/SignIn'
+import { Profile } from '@/components/Profile'
+
+import { isUserAuthenticated } from '@/utils/auth'
+
 const roboto = Roboto({
   subsets: ['latin'],
   variable: '--font-roboto',
@@ -22,6 +29,8 @@ export const metadata = {
 }
 
 export default function RootLayout({ children }: { children: ReactNode }) {
+  const isAuthenticated = isUserAuthenticated()
+
   return (
     <html lang="pt-br">
       <body
@@ -30,7 +39,29 @@ export default function RootLayout({ children }: { children: ReactNode }) {
           bg-gray-900 font-sans text-gray-100
         `}
       >
-        {children}
+        <main
+          className="
+        grid min-h-screen grid-cols-2
+      "
+        >
+          {/* Left */}
+          <div className="relative flex flex-col items-start justify-between overflow-hidden border-r border-white/10 bg-[url(../assets/bg-stars.svg)] bg-cover px-28 py-16">
+            {/* Blur */}
+            <div className="absolute right-0 top-1/2 h-[288px] w-[526px] -translate-y-1/2 translate-x-1/2 rounded-full bg-purple-700 opacity-50 blur-full" />
+
+            {/* Stripes */}
+            <div className="absolute bottom-0 right-2 top-0 w-2 bg-stripes" />
+
+            {isAuthenticated ? <Profile /> : <SignIn />}
+            <Hero />
+            <Copyright />
+          </div>
+
+          {/* Right */}
+          <div className="flex flex-col bg-[url(../assets/bg-stars.svg)] bg-cover p-16">
+            {children}
+          </div>
+        </main>
       </body>
     </html>
   )
