+++ web/src/app/memories/new/page.tsx
@@ -1,0 +1,3 @@
+export default function NewMemorie() {
+  return <h1>New</h1>
+}
